requirements.txt
langchain==0.2.16
loguru==0.7.2
astor==0.8.1
pandas==2.2.2
nltk==3.9.1
scikit-learn==1.5.1
scipy==1.14.1
statsmodels==0.14.2
langchain_openai==0.1.23
langchain_community==0.2.16
plotly==5.24.0
Pillow==10.4.0
python-dotenv==1.0.1
black==24.8.0
isort==5.13.2
pymysql==1.1.1
faiss-cpu==1.8.0.post1
kaleido==0.2.1
nbdev==2.3.29
pre-commit==3.8.0
langchain-aws==0.1.17
upstash_redis==1.1.0
fastparquet==2024.5.0
rapidfuzz==3.9.7
streamlit==1.38.0
psycopg2-binary==2.9.9
ruff==0.6.5
sqlparse==0.5.1
nbformat==5.10.4
streamlit-auth0-component==0.1.5
langchain-cohere==0.2.1
langchain-pinecone==0.1.1




